requirements.txt
Flask==3.1.0
matplotlib==3.8.4
numpy==2.2.1
pandas==2.2.3
Pillow==11.0.0
segment_anything==1.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.44.0
Werkzeug==3.1.3
